// frontend/src/api/models/dashboardSeriesResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { DashboardPeriod } from './dashboardPeriod'
import type { DashboardSeriesResponsePointsItem } from './dashboardSeriesResponsePointsItem'

export interface DashboardSeriesResponse {
  period: DashboardPeriod
  startDate: string
  endDate: string
  reportingCurrency: string
  generatedAt: string
  /** @maxItems 122 */
  points: DashboardSeriesResponsePointsItem[]
}